==> code/01_data_wrangling.Rmd <==
  ---
title: "Data wrangling"
author: "Jae Yeon Kim"
output:
  pdf_document:
    toc: yes
  html_document:
    theme: united
    toc: yes
---

# Setup 

```{r}
pacman::p_load(
  tidyverse, # for the tidyverse framework
  ggpubr, # for arranging ggplots
  ggthemes, # for fancy ggplot themes
  estimatr, # for fast estimation for design-based inference
  list, # stat methods for the item count technique and list experiment
  ggpubr, # arranging ggplots
  ggthemes, # fancy ggplot themes
  here # self-contained projects
)

devtools::install_github("jaeyk/makereproducible",
  dependencies = TRUE
)

library(makereproducible)

# Import custom functions

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)

for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme

theme_set(theme_pubr())
```

- To reduce sampling bias, I first matched the online panel with the existing probability sample (matched random sampling).
- Using this sampling frame, I contacted 2,303 people and 1,542 responded to the 1st survey (response rate: 67%).
- I randomly divided these participants into four groups (one control plus three treatment groups) blocking on ideology.
- I invited these 1,542 to the 2nd survey and 1,464 of them participated (response rate: 95%).

# Importing data

```{r}

# The entire dataset 
survey <- read.csv(make_here("/home/jae/analyzing-list-experiments/processed_data/survey_result.csv"))

```

# Wrangling data 

## Subsetting 

```{r}
# Subset those participated the pre- and post-survey = complete case analysis and drop columns that we don't need 
df <- survey %>%
  filter(answered2 == 1) %>%
  dplyr::select(-contains("ans")) %>%
    dplyr::select(-starts_with("X")) %>%
  dplyr::select(-(c("Q","time1","time2")))

```

## Renaming and recoding 

```{r}

# Check dimensions 
dim(df)

# See the names of variables 
names(df)

# Rename variables 
names(df)[which(names(df) == "pq5")] <- "partyID"
names(df)[which(names(df) == "pq7")] <- "ideology"
names(df)[which(names(df) == "z2")] <- "treat"
names(df)[which(names(df) == "q1")] <- "direct"
names(df)[which(names(df) == "q2")] <- "indirect"

df$party[df$party %in% 1] <- "Conservative Partisans"
df$party[df$party %in% 2] <- "Conservative Partisans"
df$party[df$party %in% 3] <- "Liberal Partisans"
df$party[df$party %in% 4] <- "Conservative Partisans"
df$party[df$party %in% 5] <- "Liberal Partisans"
df$party[df$party %in% 6] <- "Others"
df$party[df$party %in% 7] <- "No Party"
df$party[df$party %in% 8] <- "Don't know"

df$party_f <- factor(df$party,
  levels = c("Liberal Partisans", "Conservative Partisans", "Others", "No Party", "Don't know")
)

df$party_lib <- ifelse(df$party_f == "Liberal Partisans", 1 , 0) 
df$party_con <- ifelse(df$party_f == "Conservative Partisans", 1 , 0)

df$ideo_con <- ifelse(df$ideology == 4 | df$ideology == 5, 1, 0)
df$ideo_lib <- ifelse(df$ideology == 1 | df$ideology == 2, 1, 0)

df$college <- ifelse(df$edu1 == 4 | df$edu1 == 5, 1, 0)

df$korean_war <- ifelse(df$age >= 65, 1, 0)

df$men <- ifelse(df$gender == 1, 1, 0)

df$low_income <- ifelse(df$income %in% c(0:6), 1, 0)

df$high_income <- ifelse(df$income %in% c(7:11), 1, 0)

```

# Export the file 

```{r}

write.csv(df, make_here("/home/jae/analyzing-list-experiments/processed_data/processed_survey.csv"))

```
