==> code/02_data_analysis.Rmd <==
  ---
title: "Data analysis"
author: "Jae Yeon Kim"
output:
  pdf_document:
    toc: yes
  html_document:
    theme: united
    toc: yes
---

# Setup 

```{r}

# kable output: LaTeX
options(knitr.table.format = "latex")

pacman::p_load(
        tidyverse, # for the tidyverse framework
        ggpubr, # for arranging ggplots   
        patchwork, # for arranging ggplots
        ggthemes, # for fancy ggplot themes
        estimatr, # for fast estimation for design-based inference
        list, # statistical analysis of list experiments
        DACF, # for ceiling and or/floor data 
        MKinfer, # for inferential statistics
        nonpar, # for nonparametric inference
        here
)


library(makereproducible)

# Import custom functions

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)

for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme

theme_set(theme_pubr())
```

# Importing data

```{r}

# The processed data 
df <- read.csv(make_here("/home/jae/analyzing-list-experiments/processed_data/processed_survey.csv"))

```
  
# Testing design effect 

- Tested design effect using Bonferroni-corrected p-value. If this value is below alpha, you reject the null hypothesis of no design effect. If it is above alpha, you fail to reject the null.

- Direct bias: Above alpha. No design effect across all condition groups.
- Indirect bias: Above alpha. No design effect across all condition groups.

```{r}

design_test(df, 2)
design_test(df, 3)
design_test(df, 4)
design_test(df, 5)

```

# Multivariate analysis 

## Without concerning floor and ceiling effects 

- When ML failed: system is computationally singular

```{r}

visualize_multi_ate_group(df, 3, "ml", "North Korean refugees", -8, 5) / 
visualize_multi_ate_group(df, 2, "ml", "Low income S. Koreans", -8, 5) / 
visualize_multi_ate_group(df, 4, "ml", "Korean Chinese migrants", -8, 5) 

ggsave(here("outputs", "multi_combined.png"),
       height = 10, width = 10)

reg_table <- table_multi_ate(df, 3, "ml", "North Korean refugees", -8, 5) %>%
  filter(Name %in% c("party_lib", "ideo_lib", "income")) %>%
  group_by(Type, Condition, Name) %>%
  summarise(mean_est = round(mean(abs(Estimate)),2))
  
reg_table %>%
  arrange(mean_est) %>%
  kableExtra::kable(booktabs = TRUE)

```

# Response distributions 

```{r}

df %>%
  select(direct, indirect, treat) %>%
  rename("Direct bias" = "direct",
         "Indirect bias" = "indirect") %>%
  mutate(treat = case_when(
         treat == "1" ~ "Control",
         treat == "2" ~ "Low-income S. Koreans",
         treat == "3" ~ "North Korean refugees",
         treat == "4" ~ "Korean Chinese migrants",
         treat == "5" ~ "Indonesian migrants")) %>%
  mutate(treat = fct_relevel(treat, c("Control", "Low-income S. Koreans", "North Korean refugees", "Korean Chinese migrants", "Indonesian migrants"))) %>%
  pivot_longer(cols = c("Direct bias", "Indirect bias"),
               names_to = "Type",
               values_to = "Value") %>%
  mutate(Value = as.numeric(Value)) %>%
  ggplot(aes(x = Value, col = treat)) +
    geom_density() +
    facet_grid(~Type) +
    labs(x = "Response",
         y = "Count",
         col = "Treatment status") +
    guides(col = guide_legend(nrow = 3, byrow=TRUE))

ggsave(here("outputs", "res_dist.png"))

```

# Average treatment effect (ATE)

- By random assignment, difference-in-means is an unbiased estimator of the average treatment effect.
- Still, there's a concern about sampling variability. I address that by conducting t-tests. Here, the null hypothesis the average treatment effect is 0.

## ATE outcomes

```{r}

# Save test results
ate.results <- diff_means_test(df) 

# Recode values 
ate.results$treat <- ate.results$treat %>%
  recode(t1 = "Low-income S. Koreans",
         t2 = "North Korean refugees",
         t3 = "Korean Chinese migrants",
         t4 = "Indonesian migrants")

# Tidy model outcomes 
ate.results
```
## Visualizing results 

```{r}

ate.results %>%
  ggplot(aes(x = fct_reorder(treat, diff), y = diff, ymin = diff - conf, ymax = diff + conf)) +
  geom_pointrange() +
  geom_hline(yintercept = c(0), linetype = "dotted") +
  coord_flip() + 
  labs(x = "", y = "Estimated ATE") +
  facet_wrap(~var) +
  geom_text(aes(label = paste(round(diff,2))), position=position_dodge(width = 0.9),  vjust=-0.40)

ggsave(make_here("/home/jae/analyzing-list-experiments/outputs/ate_results_plot.png"))

````
