==> code/04_desc_survey.Rmd <==
---
title: "Descriptive stat"
author: "Jae Yeon Kim"
date: "9/4/2020"
 pdf_document:
    toc: yes
  html_document:
    theme: united
    toc: yes
---

# Setup 

```{r}

pacman::p_load(
        tidyverse, # for the tidyverse framework
        broom, 
        ggpubr, # for arranging ggplots   
        patchwork, # for arranging ggplots
        ggthemes, # for fancy ggplot themes
        here, # reproducibility
        haven # foreign data type 
)


library(makereproducible)

# Import custom functions

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)

for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme

theme_set(theme_pubr())
```

# Load files 

```{r}

# The processed data 
df <- read.csv(make_here("/home/jae/analyzing-list-experiments/processed_data/processed_survey.csv"))

```


```{r}
colnames(df)[colnames(df) == "pq5"] <- "party"

## I specified partian responses into characters.
df$party[df$party %in% 1] <- "Conservative Partisans"
df$party[df$party %in% 2] <- "Conservative Partisans"
df$party[df$party %in% 3] <- "Liberal Partisans"
df$party[df$party %in% 4] <- "Conservative Partisans"
df$party[df$party %in% 5] <- "Liberal Partisans"
df$party[df$party %in% 6] <- "Others"
df$party[df$party %in% 7] <- "No Party"
df$party[df$party %in% 8] <- "Don't know"

df$party_f <- factor(df$party,
  levels = c("Liberal Partisans", "Conservative Partisans", "Others", "No Party", "Don't know")
)


table(df$pq6[df$party_f == "Lib Party"])
table(df$pq6[df$party_f == "Con Party"])
267 / (267 + 207)
71 / (71 + 57)

# what's pk?
df$pk1 <- ifelse(df$pq8 == 1, 1, 0)
df$pk2 <- ifelse(df$pq9 == 1, 1, 0)
df$pk3 <- ifelse(df$pq10 == 3, 1, 0)

df$pk <- (df$pk1 + df$pk2 + df$pk3)
```

# Regression analysis

## Select variables
```{r}
df$external_motivation <- rowMeans(df[,c("q3","q4","q5")],na.rm = T) 
df$internal_motivation <- rowMeans(df[,c("q7","q8")],na.rm = T) 

df$lib_party <- ifelse(df$party_f == "Liberal Partisans", 1, 0)
df$conserv_party <- ifelse(df$party_f == "Conservative Partisans", 1, 0)

df %>%
  dplyr::select(q3, q4, q5) %>%
  psych::pairs.panels()

df %>%
  dplyr::select(q6, q7, q8) %>%
  psych::pairs.panels()

```

## Building models
```{r}
# Ideology 
external_reg_ideo <- lm(data = df, external_motivation ~ ideo_con + ideo_lib + income + men + college + age)

internal_reg_ideo <- lm(data = df, internal_motivation ~ ideo_con + ideo_lib + income + men + college + age)

# Partisanship 
external_reg_par <- lm(data = df, external_motivation ~ party_con + party_lib + income + men + college + age)

internal_reg_par <- lm(data = df, internal_motivation ~ party_con + party_lib + income + men + college + age)

# Visualize
ideo_plot <- bind_visualize(external_reg_ideo, internal_reg_ideo) +
  labs(title = "Ideology") +
  scale_x_discrete(labels = c("College", "Income", "Age", "Liberal", "Conservative", "Male"))

party_plot <- bind_visualize(external_reg_par, internal_reg_par) +
  labs(title = "Partisanship") +
  scale_x_discrete(labels = c("College", "Income", "Age", "Liberal", "Conservative", "Male"))

ideo_plot + party_plot 
ggsave(here("outputs", "motivations.png"))

```
